+++ src/experiments/notificationbar/implementation.js
@@ -7,14 +7,15 @@ var { EventEmitter, EventManager, ExtensionAPI } = ExtensionCommon
 var { ExtensionError } = ExtensionUtils
 var { Services } = ChromeUtils.import('resource://gre/modules/Services.jsm')
 
-class Notification {
+class ExtensionNotification {
     constructor(notificationId, properties, parent) {
         this.closedByUser = true
         this.properties = properties
         this.parent = parent
         this.notificationId = notificationId
+        this.tbVersion = this.getThunderbirdVersion().major
 
-        const { buttons, icon, label, priority, style, windowId } = properties
+        const { buttons, icon, label, priority, style, windowId, badges, placement } = properties
 
         const iconURL = icon && !icon.includes(':') ? parent.extension.baseURI.resolve(icon) : null
         const fontURL = parent.extension.baseURI.resolve('fonts/Overpass-Regular.tff')
@@ -55,7 +56,7 @@ class Notification {
         }
 
         let element
-        if (this.getThunderbirdVersion().major < 94) {
+        if (this.tbVersion < 94) {
             element = this.getNotificationBox().appendNotification(
                 label,
                 `extension-notification-${notificationId}`,
@@ -71,11 +72,40 @@ class Notification {
                     label,
                     image: iconURL,
                     priority,
+                    eventCallback: notificationBarCallback,
                 },
-                buttonSet,
-                notificationBarCallback
+                buttonSet
             )
         }
+
+        if (!element) return
+
+        if (badges.length > 0) {
+            element.removeAttribute('dismissable')
+            if (placement !== 'message') element.removeAttribute('message-bar-type')
+        }
+
+        const shadowroot = element.shadowRoot
+        const document = element.ownerDocument
+        const message = shadowroot.querySelector('label.notification-message')
+
+        badges.forEach((b) => {
+            const newDiv = document.createElement('div')
+            newDiv.classList.add('badge')
+
+            const iconURL = badges && parent.extension.baseURI.resolve(`icons/${b.type}.svg`)
+            const img = document.createElement('img')
+            img.src = iconURL
+            img.height = 20
+            img.width = 20
+
+            newDiv.appendChild(img)
+            const label = document.createElement('label')
+            label.innerText = b.value
+            newDiv.appendChild(label)
+
+            message.parentNode.insertBefore(newDiv, message.nextSibling)
+        })
 
         if (style) {
             const s = element.ownerDocument.createElement('style')
@@ -93,12 +123,22 @@ class Notification {
                     border: unset;               
                     padding: 0 1.5rem;
                 }
+                .notification-message {
+                    flex-grow: unset;
+                    flex-wrap: wrap;
+                }
                 .notification-button.small-button:hover {
                     background-color: white;
                 }
                 .infobar > .icon {
-                    width: 18px;
-                    height: 18px; 
+                    width: 24px;
+                    height: 24px; 
+                    ${badges.length > 0 && placement === 'message' ? 'display: none;' : ''}
+                }
+                .container {
+                    border-radius: ${
+                        badges.length > 0 && placement !== 'message' ? '0' : 'inherit'
+                    };
                 }
                 .infobar p {
                     font-family: 'Overpass';
@@ -108,7 +148,20 @@ class Notification {
                     --message-bar-icon-url: url(${iconURL});
                     --message-bar-text-color: ${style['color']};
                     --message-bar-background-color: ${style['background-color']};
-                }`
+                }
+                .badge {
+                    display: flex;
+                    justify-items: center;
+                    align-items: center;
+                    justify
+                    border: 0px;
+                    border-radius: 15px;
+                    background-color: #017aff;
+                    padding: 0 0.5em;
+                    margin: 0 0.25em;
+                    gap: 0.25em;
+                }
+        `
             element.shadowRoot.appendChild(s)
         }
     }
@@ -124,12 +177,42 @@ class Notification {
         }
     }
 
-    getNotificationBox() {
+    getNotificationBoxPreSupernova() {
         const w = this.parent.extension.windowManager.get(
             this.properties.windowId,
             this.parent.context
         ).window
         switch (this.properties.placement) {
+            case 'top':
+                if (!w.gExtensionNotificationTopBox) {
+                    // try to add it before the toolbox, if that fails add it firstmost
+                    const toolbox = w.document.querySelector('toolbox')
+                    if (toolbox) {
+                        w.gExtensionNotificationTopBox = new w.MozElements.NotificationBox(
+                            (element) => {
+                                element.id = 'extension-notification-top-box'
+                                element.setAttribute('notificationside', 'top')
+                                toolbox.parentElement.insertBefore(
+                                    element,
+                                    toolbox.nextElementSibling
+                                )
+                            }
+                        )
+                    } else {
+                        w.gExtensionNotificationTopBox = new w.MozElements.NotificationBox(
+                            (element) => {
+                                element.id = 'extension-notification-top-box'
+                                element.setAttribute('notificationside', 'top')
+                                w.document.documentElement.insertBefore(
+                                    element,
+                                    w.document.documentElement.firstChild
+                                )
+                            }
+                        )
+                    }
+                }
+                return w.gExtensionNotificationTopBox
+
             case 'message':
                 // below the receipient list in the message preview window
                 if (w.gMessageNotificationBar) {
@@ -164,20 +247,34 @@ class Notification {
                     )
                 }
                 return w.gExtensionNotificationBottomBox
+        }
+    }
 
+    getNotificationBoxPostSupernova() {
+        const w = this.parent.extension.windowManager.get(
+            this.properties.windowId,
+            this.parent.context
+        ).window
+        switch (this.properties.placement) {
             case 'top':
                 if (!w.gExtensionNotificationTopBox) {
-                    // try to add it before the toolbox, if that fails add it firstmost
+                    const messengerBody = w.document.getElementById('messengerBody')
                     const toolbox = w.document.querySelector('toolbox')
-                    if (toolbox) {
+                    if (messengerBody) {
                         w.gExtensionNotificationTopBox = new w.MozElements.NotificationBox(
                             (element) => {
                                 element.id = 'extension-notification-top-box'
                                 element.setAttribute('notificationside', 'top')
-                                toolbox.parentElement.insertBefore(
-                                    element,
-                                    toolbox.nextElementSibling
-                                )
+                                messengerBody.insertBefore(element, messengerBody.firstChild)
+                            }
+                        )
+                    } else if (toolbox) {
+                        w.gExtensionNotificationTopBox = new w.MozElements.NotificationBox(
+                            (element) => {
+                                element.id = 'extension-notification-top-box'
+                                element.setAttribute('notificationside', 'top')
+                                element.style.marginInlineStart = 'var(--spaces-total-width)'
+                                toolbox.insertAdjacentElement('afterend', element)
                             }
                         )
                     } else {
@@ -194,7 +291,47 @@ class Notification {
                     }
                 }
                 return w.gExtensionNotificationTopBox
+
+            case 'message': {
+                if (!this.properties.tabId) {
+                    throw new Error('appendNotification - missing tab id')
+                }
+                const aTab = this.parent.context.extension.tabManager.get(this.properties.tabId)
+                let messageBrowser = null
+                switch (aTab.nativeTab.mode?.name) {
+                    case 'mailMessageTab':
+                        // message tab;
+                        messageBrowser = aTab.nativeTab.chromeBrowser.contentWindow
+                        break
+                    case 'mail3PaneTab':
+                        // message in mail3pane tab
+                        messageBrowser =
+                            aTab.nativeTab.chromeBrowser.contentWindow.messageBrowser.contentWindow
+                        break
+                    default:
+                        // message window;
+                        messageBrowser = aTab.nativeTab.messageBrowser.contentWindow
+                        break
+                }
+                if (messageBrowser) {
+                    return messageBrowser.gMessageNotificationBar.msgNotificationBar
+                }
+                console.error(
+                    'appendNotification - could not get window for tabId ' + this.properties.tabId
+                )
+                return null
+            }
+            default:
+            case 'bottom':
+                return this.getNotificationBoxPreSupernova()
         }
+    }
+
+    getNotificationBox() {
+        if (this.tbVersion >= 112) {
+            return this.getNotificationBoxPostSupernova()
+        }
+        return this.getNotificationBoxPreSupernova()
     }
 
     remove(closedByUser) {
@@ -251,7 +388,7 @@ var notificationbar = class extends ExtensionAPI {
                     const notificationId = self.nextId++
                     self.notificationsMap.set(
                         notificationId,
-                        new Notification(notificationId, properties, self)
+                        new ExtensionNotification(notificationId, properties, self)
                     )
                     return notificationId
                 },